LeanAnimationCreated
But its broken and has the same issues that the rifle animation had:
Look down and you will be pushed up into the air, also when you do lean, it shoves you up two units. Presumably this is because the animation is reseting to the world position of the parent, unsure why this reset moves it there, if the parent should be on the ground. At least its not falling through the floor.