Fixed Animation Bug for Lean
Animation Bug for Lean:
Mistaken value was recorded in animations for lean(Left and Right) for the transform position of the object. The animation was set to move alter the transform position of the attached object on the y axis by 2. This moved the object up by 2 world units, undesirably. It is now fixed and the player can lean. Can probably do the same method for the pistol. Some touch up work might be necessary for all animations, to play a secondary animation to smoothly transition back to the start position.